msrmon: centralize subcommand required-option handling
Centralize checking of global options which are required for some
subcommands but not others, instead of checking them in each subcommand.
This avoids some duplication and eliminates the need to expose
argparse's error-reporting methods to each subcommand. It also has the
benefit of making it easier to report all such missing options at once.