msrmon: distinguish event/vehicle questions in DB
Questions all have UIDs, and for registration and change tracking, we
don't really care about the distinction between vehicle questions
(attached to assignments) and event questions (attached to attendees).
However, in order to make the --vehicle-questions and --event-questions
options work correctly, we still need to keep track of the difference.