msrmon: show questions in CSV
When --event-questions and/or --vehicle-questions are used in
combination with --csv, include all questions of the specified type in
the CSV output, after the explicitly-defined (or default) columns.
The config options csvEventQuestions and csvVehicleQuestions can be used
to control the order of question columns when these are sent; any
questions listed (by header, not UID) will be shown first and in the
specified order. Any remaining questions of the specified type will be
shown in an undefined order.

Unfortunately, we cannot default to showing questions in the order used
by MSR's CSV export, which appears to be the order in which the
questions were defined in MSR. There is no API for getting a list of all
defined questions, so we have no way to get that order. This also means
that we can only include columns for questions answered in at least one
of the entries shown, since otherwise we don't even know the question
exists.