msrmon: allow merging preloaded member/attendee info
In msr_get_{member,attendee}_info, in addition to the existing caching
mechanism, allow the caller to provide an already-loaded record to be
used as the source for the data to be merged. This will be used during
change processing to allow us to fetch the required data from MSR before
starting a database transaction, but defer merging until we've read the
old data from the database.