msrmon: sqlite timestamp converter
The sqltie3 module's adapter for datetime objects uses isoformat, which
includes a time zone offset when handling an "aware" datetime object.
The corresponding converter, however, ignores the offset and always
constructs "naive" objects with no timezone information. That's OK for
our purposes, since we always use UTC anyway.

Unfortunately, sometimes the formatted timestamp has no microseconds
part (perhaps this happens when it happens to fall on a whole second?).
But the converter can't handle a timestamp with a time zone offset but
no microseconds.

Install our own timestamp converter, which is a copy of the standard
one, but with a fix to parse out the time zone offset. We still ignore
the offset and produce naive objects, but we don't get exceptions when
trying to parse a timestamp with no microseconds.