msrmon: handle drop of non-existent assignment
When processing changes over a large period, it's possible we'll see
assignments that are added and dropped in the same changeset. Since
the assignment no longer exists in MSR, it does not get added to our
database, and so cannot be updated when the drop is processed. Don't
attempt the database update in this case.