reg_diff, reg_apply: allow group+class+car as key
Add a new -g option to reg_diff and reg_apply, which builds the row key
from the group, class, and car number, rather than just the class and
car or member number. This is necessary when the same driver/car may be
entered (in the same class!) in more than one run group.