reg_check: don't complain about blank color
We get lots of blank colors in the runoffs entries these days, because
official color is a vehicle question (so not on the main entry form)
and nothing forces (or really, even prompts) competitors to enter it.

Fortunately, for most entries we can figure out official colors by
applying heuristics to the vehicle's normal, non-runoffs color. That's
done as a separate step after reg_check runs.

Ideally, there would be configuration as to which fields are required
and thus should generate warnings if left empty. Someday.